layout relocating
