+++ src/pages/my-files.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import { graphql } from 'gatsby';
-import Layout from '../components/layout';
+import Layout from '../utils/layout';
 
 export default ({ data }) => {
   console.log(data);
   return (
     <Layout>
-      <div style={{backgroundColor: 'white', overflow: 'scroll'}}>
+      <div style={{ backgroundColor: 'white', overflow: 'scroll' }}>
         <h1>My Site's Files</h1>
         <table>
           <thead>
